Morello: change public syscall interface
Change return value data type to preserve provenence if
capability type is returned by kernel.

Add syscall test to check if syscall interface works
with capabilities as intended or not.

+++ include/unistd.h
@@ -173,7 +173,7 @@ void setusershell(void);
 void endusershell(void);
 char *getusershell(void);
 int acct(const char *);
-long syscall(long, ...);
+intptr_t syscall(long, ...);
 int execvpe(const char *, char *const [], char *const []);
 int issetugid(void);
 int getentropy(void *, size_t);
