+++ src/misc/syscall.c
@@ -5,7 +5,7 @@
 
 #undef syscall
 
-long syscall(long n, ...)
+intptr_t syscall(long n, ...)
 {
 	va_list ap;
 	syscall_arg_t a,b,c,d,e,f;
